+++ packages/sdk-coverage-tests/src/generate/emit.js
@@ -1,9 +1,17 @@
-const {useEmitter, withHistory} = require('../emitter')
-const {describeTest} = require('./describe')
+const {useEmitter} = require('../emitter')
 const {isFunction} = require('../common-util')
 
 function emitTests(tests, config) {
-  return tests.map(test => emitTest(test, config))
+  const emittedTests = []
+  const errors = []
+  for (const test of tests) {
+    try {
+      emittedTests.push(emitTest(test, config))
+    } catch (error) {
+      errors.push({test, error})
+    }
+  }
+  return {emittedTests, errors}
 }
 
 function emitTest(test, {makeSdk, fileTemplate}) {
@@ -11,19 +19,17 @@ function emitTest(test, {makeSdk, fileTemplate}) {
     throw new Error(`Missing implementation for test ${test.name}`)
   }
   test.config.baselineName = test.config.baselineName || test.key
-  const [output, emitter] = useEmitter()
-  test.output = output
   test.meta = {features: test.features}
   if (test.env) {
     test.meta.browser = test.env.browser
     test.meta.mobile = Boolean(test.env.device)
     test.meta.native = Boolean(test.env.device && !test.env.browser)
   }
-  const [history, sdk] = withHistory(makeSdk(emitter, test))
-  test.history = history
+  const [output, emitter] = useEmitter()
+  const sdk = makeSdk(emitter, test)
+  test.output = output
   if (test.page) sdk.driver.visit(test.page)
   test.test(sdk)
-  test.description = describeTest(test)
   test.code = fileTemplate(test)
   return test
 }