+++ packages/sdk-shared/scripts/render.js
@@ -85,6 +85,10 @@ const checkConfig = {
     describe: 'whether to check target fully',
     default: false,
   },
+  bcsHook: {
+    type: 'string',
+    describe: 'before capture screenshot hook',
+  }
 }
 
 const buildConfig = {
@@ -432,7 +436,7 @@ function argsToEyesConfig(args) {
     dontCloseBatches: false,
     saveNewTests: true,
     saveLogs: args.saveLogs,
-    saveDebugScreenshots: args.saveDebugScreenshots,
+    saveDebugScreenshots: args.saveDebugScreenshots
   }
 }
 
@@ -448,6 +452,7 @@ function argsToCheckConfig(args) {
     ignoreDisplacements: args.ignoreDisplacements,
     timeout: args.matchTimeout,
     isFully: args.fully,
+    hooks: args.bcsHook ? { beforeCaptureScreenshot: ['beforeCaptureScreenshot', args.bcsHook] } : undefined
   }
 }