+++ packages/eyes-nightwatch/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 
+## 1.0.3 - 2020/11/5
+
+- fix `getWindowRect` and `setWindowRect` so they work on drivers running with the JSON Wire Protocol (e.g., non-W3C)
+
 ## 1.0.2 - 2020/11/2
 
 - fix `eyesOpen` API